Enable all octo tasks to be installers in order to reduce the need to add an installer task to the start of a build definition
