+++ source/tasks/OctoCli/octocli.ts
@@ -12,14 +12,16 @@ async function run(){
         const connection = utils.getDefaultOctopusConnectionDetailsOrThrow();
         const args = tasks.getInput("args", false);
         const command = tasks.getInput("command", true);
-        const octo = utils.getOctoCommandRunner(command);
+        const octo = await utils.getOrInstallOctoCommandRunner(command);
 
         const configure = configureTool([
             connectionArguments(connection),
             includeArguments(args)
         ]);
 
-        const code:number = await configure(octo).exec();
+        const code:Number = await octo.map(configure)
+            .getOrElseL((x) => { throw new Error(x); })
+            .exec();
 
         tasks.setResult(tasks.TaskResult.Succeeded, `Succeeded executing octo command ${command} with code ${code}`);
     }catch(err){
