+++ source/tasks/Push/task.json
@@ -16,8 +16,7 @@
         "Patch": 0
     },
     "demands": [
-        "DotNetCore",
-        "Octo"
+        "DotNetCore"
     ],
     "minimumAgentVersion": "1.95.0",
     "groups": [
