+++ source/tasks/Utils/index.ts
@@ -2,3 +2,4 @@ export * from "./connection";
 export * from "./tool";
 export * from "./inputs";
 export * from "./environment";
+export * from "./install";
